+++ extensions/elation/actions/addAllergy/config/dataPoints.ts
@@ -1,0 +1,8 @@
+import { type DataPointDefinition } from '@awell-health/extensions-core'
+
+export const dataPoints = {
+  allergyId: {
+    key: 'allergyId',
+    valueType: 'number',
+  },
+} satisfies Record<string, DataPointDefinition>
